fix(tests): Rework test handling, starting/stopping ProcessEngines. (#802)
Removed extra caching mechanism in TestHelper and rely on caches in
ProcessEngines only (avoiding caches containing different
ProcessEngine's with same name).

Refactor ProcessEngineExtension to start ProcessEngine on start of the
test and not when calling "build". This moves the initialization of the
ProcessEngine to happen right before the start of the test. In the past
it was when the class was loaded (static initializier), causing
flickering tests in different runtime scenarios (Maven vs. IDE). Adapted
other Extensions accordingly.

Made ProcessEngines thread save by using "volatile" and thread safe
Collections. This ensures that worker Threads are guaranteed to see
updates made to the ProcesssEngineConfiguration. Also used "volatile" on
fields of the ProcessEngineConfigurationImpl as well for same reason.

Adapt ProcessEngineBootstrapRule to generate ProcessEngine's with unique
names to ensure they do not collide or accidently close the "default"
engine.

Adapt tests to include names for ProcessEngine's to ensure they do not
accidently use the "default" ProcessEngine (except for the tests that
actually SHOULD use the "default" engine, of course).

Add "static" to ProcessEngineExtension fields where it is missing. Make
general usage of modifiers more consistent.

Provide additional builder methods for ProcessEngineExtension for process engine name randomization and auto-closing after test execution.

---------

+++ engine/src/test/java/org/operaton/bpm/engine/test/api/identity/IdentityServiceTenantTest.java
@@ -19,6 +19,7 @@ package org.operaton.bpm.engine.test.api.identity;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.Assertions.fail;
+import static org.operaton.bpm.engine.test.util.ProcessEngineUtils.newRandomProcessEngineName;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
@@ -51,6 +52,7 @@ class IdentityServiceTenantTest {
   protected static final String TENANT_TWO = "tenant2";
 
   private static final String INVALID_ID_MESSAGE = "%s has an invalid id: '%s' is not a valid resource identifier.";
+  private static final String PROCESS_ENGINE_NAME = newRandomProcessEngineName();
 
   protected IdentityService identityService;
   protected ProcessEngine processEngine;
@@ -66,7 +68,7 @@ class IdentityServiceTenantTest {
     identityService.deleteUser(USER_ONE);
     identityService.deleteUser(USER_TWO);
 
-    if (processEngine != null) {
+    if (processEngine != null && ProcessEngines.getDefaultProcessEngine() != processEngine) {
       for (Tenant deleteTenant : processEngine.getIdentityService().createTenantQuery().list()) {
         processEngine.getIdentityService().deleteTenant(deleteTenant.getId());
       }
@@ -159,6 +161,7 @@ class IdentityServiceTenantTest {
   void testCustomCreateTenantWhitelistPattern() {
     processEngine = ProcessEngineConfiguration
       .createProcessEngineConfigurationFromResource("org/operaton/bpm/engine/test/api/identity/generic.resource.id.whitelist.operaton.cfg.xml")
+      .setProcessEngineName(PROCESS_ENGINE_NAME)
       .buildProcessEngine();
     processEngine.getProcessEngineConfiguration().setTenantResourceWhitelistPattern("[a-zA-Z]+");
 
@@ -177,6 +180,7 @@ class IdentityServiceTenantTest {
   void testCustomTenantWhitelistPattern() {
     processEngine = ProcessEngineConfiguration
       .createProcessEngineConfigurationFromResource("org/operaton/bpm/engine/test/api/identity/generic.resource.id.whitelist.operaton.cfg.xml")
+      .setProcessEngineName(PROCESS_ENGINE_NAME)
       .buildProcessEngine();
     processEngine.getProcessEngineConfiguration().setTenantResourceWhitelistPattern("[a-zA-Z]+");
 
@@ -234,6 +238,7 @@ class IdentityServiceTenantTest {
   void createTenantWithGenericResourceId() {
     processEngine = ProcessEngineConfiguration
       .createProcessEngineConfigurationFromResource("org/operaton/bpm/engine/test/api/identity/generic.resource.id.whitelist.operaton.cfg.xml")
+      .setProcessEngineName(PROCESS_ENGINE_NAME)
       .buildProcessEngine();
 
     Tenant tenant = processEngine.getIdentityService().newTenant("*");
